Keep shortcut ref links
